Add the world...
Ok, this has a lot of binaries in it. This is intentional, since it gives a snapshot of the current "build" along with the source for it.

This code is updated to build from Eclipse, it probably won't work with gradle any more due to changed paths.

The functional changes are as follows;
Strip out registration/donation parts,
Stript out google analytics parts,
Add the ability to *completely* block wakelocks and alarms by inputting a negative value in the time setting.